Updating model to use float for points

+++ model/team_result.rb
@@ -3,7 +3,7 @@ require 'mongo_mapper'
 class TeamResult
 	include MongoMapper::Document
 
-	key :points, Integer
+	key :points, Float
 
 	belongs_to :matchup
 	belongs_to :user
